se movió la logica a logica_pantalla y refactorizó las validaciones
