+++ base_datos/centrosacopio.txt
@@ -17,3 +17,4 @@ tecnolandoia|12345678|dsada
 tecnolandoia|11111111|dfsdf
 tecnolandoia|12345678|sdfsdf
 tecnolandoia|77777777|gsfsdf
+Sprint 2|12345678|sfsdf
