+++ base_datos/sedes.txt
@@ -1,2 +1,4 @@
 tecnolandoia|Alajuela|10203040|activo
 Sprint 2|Alajuela|88888888|activo
+TECSANJOSE|San José|12345678|activo
+TecSanCarlosMiLindaTierra|Cartago|12345678|activo
