+++ pantallas/__init__.py
@@ -1,10 +1,4 @@
 from .sedes_pntlla import sedes_pantalla
-from .logica_sedes_pntlla import crear_sede,limpiar_campos
-
 from .centrosacopio_pntlla import centrosacopio_panatlla
-from .logica_centrosacopio_pntlla import crear_centro_acopio
-
-from .logica_catalogo_materiales_pntlla import crear_material,actualizar_lista
 from .catalogo_materiales_pntlla import catalogo_materiales_pantalla
-
 from .crear_transaccion_pntlla import crear_transaccion_pntlla
