+++ pantallas_logica/__init__.py
@@ -1,0 +1,4 @@
+
+from .catalogo_materiales_logica_pntlla import crear_material,actualizar_lista
+from .centrosacopio_logica_pntlla import crear_centro_acopio
+from .sedes_logica_pntlla import crear_sede,limpiar_campos
